Changes in quality, progress in QTable Quality is now a float, not an int. Implemented a few functions in QTable

+++ LeaveLabyrinth/Assets/Scripts/ActionQuality.cs
@@ -8,9 +8,9 @@ public class ActionQuality
 {
 	public int m_ActionID { get; set; }
 
-	public int m_Quality{ get; set; }
+	public float m_Quality{ get; set; }
 
-	public ActionQuality (int actionID, int quality)
+	public ActionQuality (int actionID, float quality)
 	{
 		m_ActionID = actionID;
 		m_Quality = quality;
